Add files via upload
Linear regression is a supervised machine learning method that is used by the Train Using AutoML tool and finds a linear equation that best describes the correlation of the explanatory variables with the dependent variable. This is achieved by fitting a line to the data using least squares. performs the task to predict a dependent variable value (y) based on a given independent variable (x))predicting the price of a house by analyzing sales data of that region